requirements.txt
google-generativeai
pandas
scikit-learn
joblib
openpyxl
